+++ packages/custom-survey-client/src/index.ts
@@ -11,6 +11,7 @@ export * from "./SurveyQualityPanel";
 export * from "./SurveyVersionHistory";
 export * from "./SurveyVersionPanel";
 export * from "./shared";
+export * from "./translation";
 export * from "./types";
 export * from "./versionOperations";
 export * from "./workflow";