+++ packages/custom-survey-client/src/quality/index.ts
@@ -1,2 +1,3 @@
 export * from "./controller";
+export * from "./mapper";
 export * from "./types";